Refactor reusable parts of CommentSection into CommentList
