+++ src/components/common/CommentList/CommentList.css
@@ -1,0 +1,10 @@
+.commentList {
+    display: flex;
+    justify-content: center;
+}
+
+ul {
+    list-style-type: none;
+    margin: 0;
+    padding: 0;
+}
