+++ src/components/features/CommentSection/CommentSection.css
@@ -1,10 +1,1 @@
-.commentSection {
-    display: flex;
-    justify-content: center;
-}
 
-ul {
-    list-style-type: none;
-    margin: 0;
-    padding: 0;
-}
